+++ blocks/bs-card/editor.js
@@ -121,6 +121,18 @@
             },
             animationMobileEnabled: {
                 type: 'boolean'
+            },
+            animationScrollStart: {
+                type: 'string',
+                default: 'top 70%'
+            },
+            animationScrollEnd: {
+                type: 'string',
+                default: 'top 10%'
+            },
+            animationScrollMarkers: {
+                type: 'boolean',
+                default: false
             }
         },