+++ inc/template-tags.php
@@ -99,7 +99,10 @@ function ileben_render_loader()
     ?>
     <div id="site-loader" class="site-loader" aria-hidden="true">
         <div class="loader-inner text-center">
-            <img src="<?php echo ILEBEN_THEME_URI . '/assets/images/logo-leben-solo.svg'; ?>" alt="<?php esc_attr_e('Ileben', 'ileben-landing'); ?>" class="img-logo">
+            <img src="<?php echo ILEBEN_THEME_URI . '/assets/images/logo-leben-solo.svg'; ?>" alt="<?php esc_attr_e('Ileben', 'ileben-landing'); ?>" 
+            class="img-logo" 
+            style="width: auto; margin-bottom: 1rem; height: 7dvh;"
+            >
             <br>
             <div class="spinner-border text-primary" role="status">
                 <span class="visually-hidden">Loading...</span>